pyproject.toml
[project]
name = "aws-orchestrator-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "a2a-sdk>=0.2.9",
    "chromadb>=1.0.15",
    "colorama>=0.4.6",
    "langchain>=0.3.26",
    "langchain-community>=0.3.26",
    "langchain-core>=0.3.66",
    "langchain-openai>=0.3.27",
    "langchain-perplexity>=0.1.1",
    "langgraph>=0.5.0",
    "langgraph-supervisor>=0.0.28",
    "loguru>=0.7.3",
    "pinecone>=7.3.0",
    "pydantic>=2.11.7",
    "pytest>=8.4.1",
    "pytest-asyncio>=0.24.0",
    "uvicorn>=0.35.0",
    "websockets>=15.0.1",
    "click>=8.0.0",
    "httpx>=0.24.0",
    "asyncclick>=8.1.8",
    "langchain-mcp-adapters>=0.1.9",
    "langchain-tavily>=0.2.11",
    "json-repair>=0.50.1",
    "langgraph-swarm>=0.0.13",
]

[project.scripts]
aws-orchestrator-agent = "aws_orchestrator_agent.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["aws_orchestrator_agent"]
